requirements.txt
# /requirements.txt
# title: Pythonパッケージ依存関係リスト
# role: このプロジェクトの実行に必要なすべてのPythonライブラリとそのバージョンを定義する。

# --- 主要AI・LLM関連ライブラリ ---
ollama
langchain
langchain-core
langchain-community
langchain-google-community

# --- ベクトルストア・検索関連 ---
faiss-cpu
wikipedia
google-search-results

# --- アプリケーションフレームワーク・ユーティリティ ---
dependency-injector
pydantic

# --- 数値計算ライブラリ (バージョン指定) ---
numpy<2.0

# --- 環境変数 ---
python-dotenv